requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-dotenv==1.0.1
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
